Updated to CodeAnalyzer version 3.8.0-3.final
Updated to newest version of the Microsoft.CodeAnalyzer package and
used new names for GeneratorContexts.

My old approach for the unit tests which used reflection does not work
any more and, well, it relied on reflection. Therefore, I am now using
the CSharpGeneratorDriver class, which makes the tests much better.